requirements.txt
python-dotenv
pymongo
streamlit
